refactor (step4) : 클래스 별 패키지 구분
