+++ src/main/java/lotto/domain/ticket/LottoTicket.java
@@ -1,4 +1,6 @@
-package lotto;
+package lotto.domain.ticket;
+
+import lotto.domain.LottoNumber;
 
 import java.util.*;
 import java.util.stream.Collectors;
