+++ src/main/java/lotto/domain/ticket/LottoTicketAutoGenerator.java
@@ -1,4 +1,6 @@
-package lotto;
+package lotto.domain.ticket;
+
+import lotto.domain.LottoNumber;
 
 import java.util.Collections;
 import java.util.HashSet;
