+++ src/test/java/lotto/LottoAnalyzerTest.java
@@ -1,6 +1,7 @@
 package lotto;
 
-import org.assertj.core.util.Sets;
+import lotto.domain.*;
+import lotto.domain.ticket.LottoTicket;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -8,7 +9,6 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
 import java.util.Arrays;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
