+++ src/test/java/lotto/LottoRankTest.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.domain.LottoRank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
