+++ src/test/java/lotto/ticket/LottoTicketGeneratorTest.java
@@ -1,5 +1,7 @@
-package lotto;
+package lotto.ticket;
 
+import lotto.domain.ticket.LottoTicket;
+import lotto.domain.ticket.LottoTicketAutoGenerator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
